+++ resources/lang/nl/versioning.php
@@ -7,6 +7,7 @@ return [
     'file_replace_failed' => 'Het bestand kon niet vervangen worden',
     'unsupported_extension' => 'Bestanden met de extensie :extension worden niet ondersteund',
     'version_reverted' => 'Teruggezet naar versie :version',
+    'version_files_missing' => 'Deze versie kan niet meer teruggezet worden, de bestanden zijn niet meer opgeslagen',
     'revert_confirm_heading' => 'Ben je zeker dat je naar deze versie wil terugzetten?',
     'no_versions' => 'Geen eerdere versies',
 ];